move ffi and fap out
