+++ new-mock-engine-store/src/mock_store/mock_engine_store_server.rs
@@ -1,17 +1,10 @@
 // Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::{
-    cell::RefCell,
-    pin::Pin,
-    sync::{atomic::Ordering, Mutex},
-    time::Duration,
-};
-
-use int_enum::IntEnum;
+use std::{cell::RefCell, pin::Pin, sync::atomic::Ordering};
 
 use crate::{
     common::*, interfaces_ffi::PageAndCppStrWithView, mock_cluster, mock_cluster::TiFlashEngine,
-    mock_core::*, mock_page_storage::*, node::NodeCluster, server::ServerCluster,
+    mock_core::*, mock_fast_add_peer::*, mock_ffi::*, mock_page_storage::*, server::ServerCluster,
 };
 
 pub struct EngineStoreServer {
@@ -109,62 +102,7 @@ pub struct EngineStoreServerWrap {
     pub cluster_ptr: isize,
 }
 
-fn set_new_region_peer(new_region: &mut MockRegion, store_id: u64) {
-    if let Some(peer) = new_region
-        .region
-        .get_peers()
-        .iter()
-        .find(|&peer| peer.get_store_id() == store_id)
-    {
-        new_region.peer = peer.clone();
-    } else {
-        // This happens when region is not found.
-    }
-}
-
-pub fn make_new_region(
-    maybe_from_region: Option<kvproto::metapb::Region>,
-    maybe_store_id: Option<u64>,
-) -> MockRegion {
-    let mut region = MockRegion {
-        region: maybe_from_region.unwrap_or_default(),
-        ..Default::default()
-    };
-    if let Some(store_id) = maybe_store_id {
-        set_new_region_peer(&mut region, store_id);
-    }
-    region
-        .apply_state
-        .mut_truncated_state()
-        .set_index(raftstore::store::RAFT_INIT_LOG_INDEX);
-    region
-        .apply_state
-        .mut_truncated_state()
-        .set_term(raftstore::store::RAFT_INIT_LOG_TERM);
-    region.set_applied(
-        raftstore::store::RAFT_INIT_LOG_INDEX,
-        raftstore::store::RAFT_INIT_LOG_TERM,
-    );
-    region
-}
-
-pub fn write_kv_in_mem(region: &mut MockRegion, cf_index: usize, k: &[u8], v: &[u8]) {
-    let data = &mut region.data[cf_index];
-    let pending_delete = &mut region.pending_delete[cf_index];
-    let pending_write = &mut region.pending_write[cf_index];
-    pending_delete.remove(k);
-    data.insert(k.to_vec(), v.to_vec());
-    pending_write.insert(k.to_vec(), v.to_vec());
-}
-
-fn delete_kv_in_mem(region: &mut MockRegion, cf_index: usize, k: &[u8]) {
-    let data = &mut region.data[cf_index];
-    let pending_delete = &mut region.pending_delete[cf_index];
-    pending_delete.insert(k.to_vec());
-    data.remove(k);
-}
-
-unsafe fn load_data_from_db(store: &mut EngineStoreServer, region_id: u64) {
+pub(crate) unsafe fn load_data_from_db(store: &mut EngineStoreServer, region_id: u64) {
     let store_id = store.id;
     let engine = &mut store.engines.as_mut().unwrap().kv;
     let region = store.kvstore.get_mut(&region_id).unwrap();
@@ -248,42 +186,6 @@ pub(crate) unsafe fn write_to_db_data_by_engine(
             kv.rocks.delete_cf(cf_name, &tikv_key).unwrap();
         }
     }
-}
-
-pub unsafe fn create_cpp_str_parts(
-    s: Option<Vec<u8>>,
-) -> (interfaces_ffi::RawCppPtr, interfaces_ffi::BaseBuffView) {
-    match s {
-        Some(s) => {
-            let len = s.len() as u64;
-            let ptr = Box::into_raw(Box::new(s)); // leak
-            (
-                interfaces_ffi::RawCppPtr {
-                    ptr: ptr as RawVoidPtr,
-                    type_: RawCppPtrTypeImpl::String.into(),
-                },
-                interfaces_ffi::BaseBuffView {
-                    data: (*ptr).as_ptr() as *const _,
-                    len,
-                },
-            )
-        }
-        None => (
-            interfaces_ffi::RawCppPtr {
-                ptr: std::ptr::null_mut(),
-                type_: RawCppPtrTypeImpl::None.into(),
-            },
-            interfaces_ffi::BaseBuffView {
-                data: std::ptr::null(),
-                len: 0,
-            },
-        ),
-    }
-}
-
-pub unsafe fn create_cpp_str(s: Option<Vec<u8>>) -> interfaces_ffi::CppStrWithView {
-    let (p, v) = create_cpp_str_parts(s);
-    interfaces_ffi::CppStrWithView { inner: p, view: v }
 }
 
 #[allow(clippy::single_element_loop)]
@@ -781,35 +683,6 @@ unsafe extern "C" fn ffi_handle_write_raft_cmd(
     store.handle_write_raft_cmd(arg2, arg3)
 }
 
-#[repr(u32)]
-#[derive(IntEnum, Copy, Clone)]
-pub enum RawCppPtrTypeImpl {
-    None = 0,
-    String = 1,
-    PreHandledSnapshotWithBlock = 11,
-    WakerNotifier = 12,
-    PSWriteBatch = 13,
-    PSUniversalPage = 14,
-    PSPageAndCppStr = 15,
-}
-
-impl From<RawCppPtrTypeImpl> for interfaces_ffi::RawCppPtrType {
-    fn from(value: RawCppPtrTypeImpl) -> Self {
-        assert_type_eq::assert_type_eq!(interfaces_ffi::RawCppPtrType, u32);
-        value.int_value()
-    }
-}
-
-impl From<interfaces_ffi::RawCppPtrType> for RawCppPtrTypeImpl {
-    fn from(value: interfaces_ffi::RawCppPtrType) -> Self {
-        if let Ok(s) = RawCppPtrTypeImpl::from_int(value) {
-            s
-        } else {
-            panic!("unknown RawCppPtrType {:?}", value);
-        }
-    }
-}
-
 extern "C" fn ffi_need_flush_data(
     _arg1: *mut interfaces_ffi::EngineStoreServerWrap,
     _region_id: u64,
@@ -870,95 +743,6 @@ unsafe extern "C" fn ffi_try_flush_data(
         "fn_try_flush_data".to_string(),
     );
     true as u8
-}
-
-extern "C" fn ffi_gen_cpp_string(s: interfaces_ffi::BaseBuffView) -> interfaces_ffi::RawCppPtr {
-    let str = Box::new(Vec::from(s.to_slice()));
-    let ptr = Box::into_raw(str);
-    interfaces_ffi::RawCppPtr {
-        ptr: ptr as *mut _,
-        type_: RawCppPtrTypeImpl::String.into(),
-    }
-}
-
-pub struct RawCppStringPtrGuard(interfaces_ffi::RawCppStringPtr);
-
-impl Default for RawCppStringPtrGuard {
-    fn default() -> Self {
-        Self(std::ptr::null_mut())
-    }
-}
-
-impl std::convert::AsRef<interfaces_ffi::RawCppStringPtr> for RawCppStringPtrGuard {
-    fn as_ref(&self) -> &interfaces_ffi::RawCppStringPtr {
-        &self.0
-    }
-}
-impl std::convert::AsMut<interfaces_ffi::RawCppStringPtr> for RawCppStringPtrGuard {
-    fn as_mut(&mut self) -> &mut interfaces_ffi::RawCppStringPtr {
-        &mut self.0
-    }
-}
-
-impl Drop for RawCppStringPtrGuard {
-    fn drop(&mut self) {
-        interfaces_ffi::RawCppPtr {
-            ptr: self.0 as *mut _,
-            type_: RawCppPtrTypeImpl::String.into(),
-        };
-    }
-}
-
-impl RawCppStringPtrGuard {
-    pub fn as_str(&self) -> &[u8] {
-        let s = self.0 as *mut Vec<u8>;
-        unsafe { &*s }
-    }
-}
-
-pub struct ProxyNotifier {
-    cv: std::sync::Condvar,
-    mutex: Mutex<()>,
-    // multi notifiers single receiver model. use another flag to avoid waiting until timeout.
-    flag: std::sync::atomic::AtomicBool,
-}
-
-impl ProxyNotifier {
-    pub fn blocked_wait_for(&self, timeout: Duration) {
-        // if flag from false to false, wait for notification.
-        // if flag from true to false, do nothing.
-        if !self.flag.swap(false, std::sync::atomic::Ordering::AcqRel) {
-            {
-                let lock = self.mutex.lock().unwrap();
-                if !self.flag.load(std::sync::atomic::Ordering::Acquire) {
-                    let _cv = self.cv.wait_timeout(lock, timeout);
-                }
-            }
-            self.flag.store(false, std::sync::atomic::Ordering::Release);
-        }
-    }
-
-    pub fn wake(&self) {
-        // if flag from false -> true, then wake up.
-        // if flag from true -> true, do nothing.
-        if !self.flag.swap(true, std::sync::atomic::Ordering::AcqRel) {
-            let _lock = self.mutex.lock().unwrap();
-            self.cv.notify_one();
-        }
-    }
-
-    pub fn new_raw() -> RawCppPtr {
-        let notifier = Box::new(Self {
-            cv: Default::default(),
-            mutex: Mutex::new(()),
-            flag: std::sync::atomic::AtomicBool::new(false),
-        });
-
-        RawCppPtr {
-            ptr: Box::into_raw(notifier) as _,
-            type_: RawCppPtrTypeImpl::WakerNotifier.into(),
-        }
-    }
 }
 
 extern "C" fn ffi_gc_special_raw_cpp_ptr(
@@ -1325,297 +1109,4 @@ unsafe extern "C" fn ffi_handle_compute_store_stats(
         engine_bytes_read: 0,
         engine_keys_read: 0,
     }
-}
-
-pub fn copy_meta_from<EK: engine_traits::KvEngine, ER: RaftEngine + engine_traits::Peekable>(
-    source_engines: &Engines<EK, ER>,
-    target_engines: &Engines<EK, ER>,
-    source: &MockRegion,
-    target: &mut MockRegion,
-    new_region_meta: kvproto::metapb::Region,
-    copy_region_state: bool,
-    copy_apply_state: bool,
-    copy_raft_state: bool,
-) -> raftstore::Result<()> {
-    let region_id = source.region.get_id();
-
-    let mut wb = target_engines.kv.write_batch();
-
-    // Can't copy this key, otherwise will cause a bootstrap.
-    // box_try!(wb.put_msg(keys::PREPARE_BOOTSTRAP_KEY, &source.region));
-
-    // region local state
-    if copy_region_state {
-        let mut state = RegionLocalState::default();
-        state.set_region(new_region_meta);
-        box_try!(wb.put_msg_cf(CF_RAFT, &keys::region_state_key(region_id), &state));
-    }
-
-    // apply state
-    if copy_apply_state {
-        let apply_state: RaftApplyState =
-            match general_get_apply_state(&source_engines.kv, region_id) {
-                Some(x) => x,
-                None => return Err(box_err!("bad RaftApplyState")),
-            };
-        wb.put_msg_cf(CF_RAFT, &keys::apply_state_key(region_id), &apply_state)?;
-        target.apply_state = apply_state.clone();
-        target.applied_term = source.applied_term;
-    }
-
-    wb.write()?;
-    target_engines.sync_kv()?;
-
-    let mut raft_wb = target_engines.raft.log_batch(1024);
-    // raft state
-    if copy_raft_state {
-        let raft_state = match get_raft_local_state(&source_engines.raft, region_id) {
-            Some(x) => x,
-            None => return Err(box_err!("bad RaftLocalState")),
-        };
-        raft_wb.put_raft_state(region_id, &raft_state)?;
-    };
-
-    box_try!(target_engines.raft.consume(&mut raft_wb, true));
-    Ok(())
-}
-
-pub fn copy_data_from(
-    source_engines: &Engines<impl KvEngine, impl RaftEngine + engine_traits::Peekable>,
-    target_engines: &Engines<impl KvEngine, impl RaftEngine>,
-    source: &MockRegion,
-    target: &mut MockRegion,
-) -> raftstore::Result<()> {
-    let region_id = source.region.get_id();
-
-    // kv data in memory
-    for cf in 0..3 {
-        for (k, v) in &source.data[cf] {
-            // debug!("copy_data_from region {} {:?} {:?}", region_id, k, v);
-            write_kv_in_mem(target, cf, k.as_slice(), v.as_slice());
-        }
-    }
-
-    // raft log
-    let mut raft_wb = target_engines.raft.log_batch(1024);
-    let mut entries: Vec<raft::eraftpb::Entry> = Default::default();
-    source_engines
-        .raft
-        .get_all_entries_to(region_id, &mut entries)
-        .unwrap();
-    debug!("copy raft log {:?}", entries);
-
-    raft_wb.append(region_id, None, entries)?;
-    box_try!(target_engines.raft.consume(&mut raft_wb, true));
-    Ok(())
-}
-
-#[allow(clippy::redundant_closure_call)]
-unsafe extern "C" fn ffi_fast_add_peer(
-    arg1: *mut interfaces_ffi::EngineStoreServerWrap,
-    region_id: u64,
-    new_peer_id: u64,
-) -> interfaces_ffi::FastAddPeerRes {
-    let store = into_engine_store_server_wrap(arg1);
-    let cluster = &*(store.cluster_ptr as *const mock_cluster::Cluster<NodeCluster>);
-    let store_id = (*store.engine_store_server).id;
-    (*store.engine_store_server).mutate_region_states(region_id, |e: &mut RegionStats| {
-        e.fast_add_peer_count.fetch_add(1, Ordering::SeqCst);
-    });
-
-    let failed_add_peer_res =
-        |status: interfaces_ffi::FastAddPeerStatus| interfaces_ffi::FastAddPeerRes {
-            status,
-            apply_state: create_cpp_str(None),
-            region: create_cpp_str(None),
-        };
-    let from_store = (|| {
-        fail::fail_point!("fap_mock_add_peer_from_id", |t| {
-            let t = t.unwrap().parse::<u64>().unwrap();
-            t
-        });
-        1
-    })();
-    let block_wait: bool = (|| {
-        fail::fail_point!("fap_mock_block_wait", |t| {
-            let t = t.unwrap().parse::<u64>().unwrap();
-            t
-        });
-        0
-    })() != 0;
-    let fail_after_write: bool = (|| {
-        fail::fail_point!("fap_mock_fail_after_write", |t| {
-            let t = t.unwrap().parse::<u64>().unwrap();
-            t
-        });
-        0
-    })() != 0;
-    debug!("recover from remote peer: enter from {} to {}", from_store, store_id; "region_id" => region_id);
-
-    for retry in 0..300 {
-        let mut ret: Option<interfaces_ffi::FastAddPeerRes> = None;
-        if retry > 0 {
-            std::thread::sleep(std::time::Duration::from_millis(30));
-        }
-        cluster.access_ffi_helpers(&mut |guard: &mut HashMap<u64, crate::mock_cluster::FFIHelperSet>| {
-            debug!("recover from remote peer: preparing from {} to {}, persist and check source", from_store, store_id; "region_id" => region_id);
-            let source_server = match guard.get_mut(&from_store) {
-                Some(s) => &mut s.engine_store_server,
-                None => {
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::NoSuitable));
-                    return;
-                }
-            };
-            let source_engines = match source_server.engines.clone() {
-                Some(s) => s,
-                None => {
-                    error!("recover from remote peer: failed get source engine"; "region_id" => region_id);
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData));
-                    return
-                }
-            };
-            // TODO We must ask the remote peer to persist before get a snapshot.
-            let source_region = match source_server.kvstore.get(&region_id) {
-                Some(s) => s,
-                None => {
-                    error!("recover from remote peer: failed read source region info"; "region_id" => region_id);
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData));
-                    return;
-                }
-            };
-            let region_local_state: RegionLocalState = match general_get_region_local_state(
-                &source_engines.kv,
-                region_id,
-            ) {
-                Some(x) => x,
-                None => {
-                    debug!("recover from remote peer: preparing from {} to {}:{}, not region state", from_store, store_id, new_peer_id; "region_id" => region_id);
-                    // We don't return BadData here, since the data may not be persisted.
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::WaitForData));
-                    return;
-                }
-            };
-            let new_region_meta = region_local_state.get_region();
-            let peer_state = region_local_state.get_state();
-            // Validation
-            match peer_state {
-                PeerState::Tombstone | PeerState::Applying => {
-                    // Note in real implementation, we will avoid selecting this peer.
-                    error!("recover from remote peer: preparing from {} to {}:{}, error peer state {:?}", from_store, store_id, new_peer_id, peer_state; "region_id" => region_id);
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData));
-                    return;
-                }
-                _ => {
-                    info!("recover from remote peer: preparing from {} to {}:{}, ok peer state {:?}", from_store, store_id, new_peer_id, peer_state; "region_id" => region_id);
-                }
-            };
-            if !engine_store_ffi::core::validate_remote_peer_region(
-                new_region_meta,
-                store_id,
-                new_peer_id,
-            ) {
-                debug!("recover from remote peer: preparing from {} to {}, not applied conf change {}", from_store, store_id, new_peer_id; "region_id" => region_id);
-                ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::WaitForData));
-                return;
-            }
-            // TODO check commit_index and applied_index here
-            debug!("recover from remote peer: preparing from {} to {}, check target", from_store, store_id; "region_id" => region_id);
-            let new_region = make_new_region(
-                Some(new_region_meta.clone()),
-                Some((*store.engine_store_server).id),
-            );
-            (*store.engine_store_server)
-                .kvstore
-                .insert(region_id, Box::new(new_region));
-            let target_engines = match (*store.engine_store_server).engines.clone() {
-                Some(s) => s,
-                None => {
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::OtherError));
-                    return;
-                }
-            };
-            let target_region = match (*store.engine_store_server).kvstore.get_mut(&region_id) {
-                Some(s) => s,
-                None => {
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData));
-                    return;
-                }
-            };
-            debug!("recover from remote peer: meta from {} to {}", from_store, store_id; "region_id" => region_id);
-            // Must first dump meta then data, otherwise data may lag behind.
-            // We can see a raft log hole at applied_index otherwise.
-            let apply_state: RaftApplyState = match general_get_apply_state(
-                &source_engines.kv,
-                region_id,
-            ) {
-                Some(x) => x,
-                None => {
-                    error!("recover from remote peer: failed read apply state"; "region_id" => region_id);
-                    ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData));
-                    return;
-                }
-            };
-            debug!("recover from remote peer: begin data from {} to {}", from_store, store_id; 
-                "region_id" => region_id,
-                "apply_state" => ?apply_state,
-            );
-            // TODO In TiFlash we should take care of write batch size
-            if let Err(e) = copy_data_from(
-                &source_engines,
-                &target_engines,
-                &source_region,
-                target_region,
-            ) {
-                error!("recover from remote peer: inject error {:?}", e; "region_id" => region_id);
-                ret = Some(failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::FailedInject));
-                return;
-            }
-            if fail_after_write {
-                let mut raft_wb = target_engines.raft.log_batch(1024);
-                let mut entries: Vec<raft::eraftpb::Entry> = Default::default();
-                target_engines
-                    .raft
-                    .get_all_entries_to(region_id, &mut entries)
-                    .unwrap();
-                let l = entries.len();
-                // Manually delete one raft log
-                // let from = entries.get(l - 2).unwrap().get_index();
-                let from = 7;
-                let to = entries.get(l - 1).unwrap().get_index() + 1;
-                debug!("recover from remote peer: simulate error from {} to {}", from_store, store_id;
-                    "region_id" => region_id,
-                    "from" => from,
-                    "to" => to,
-                );
-                // raft_wb.cut_logs(region_id, from, to);
-                target_engines.raft.gc(region_id, from, to, &mut raft_wb).unwrap();
-                target_engines.raft.consume(&mut raft_wb, true).unwrap();
-            }
-            let apply_state_bytes = apply_state.write_to_bytes().unwrap();
-            let region_bytes = region_local_state.get_region().write_to_bytes().unwrap();
-            let apply_state_ptr = create_cpp_str(Some(apply_state_bytes));
-            let region_ptr = create_cpp_str(Some(region_bytes));
-            // Check if we have commit_index.
-            debug!("recover from remote peer: ok from {} to {}", from_store, store_id; "region_id" => region_id);
-            ret = Some(interfaces_ffi::FastAddPeerRes {
-                status: interfaces_ffi::FastAddPeerStatus::Ok,
-                apply_state: apply_state_ptr,
-                region: region_ptr,
-            });
-        });
-        if let Some(r) = ret {
-            match r.status {
-                interfaces_ffi::FastAddPeerStatus::WaitForData => {
-                    if block_wait {
-                        continue;
-                    } else {
-                        return r;
-                    }
-                }
-                _ => return r,
-            }
-        }
-    }
-    error!("recover from remote peer: failed after retry"; "region_id" => region_id);
-    failed_add_peer_res(interfaces_ffi::FastAddPeerStatus::BadData)
 }
