+++ new-mock-engine-store/src/mock_store/mod.rs
@@ -2,9 +2,10 @@
 pub(crate) mod common;
 pub mod mock_core;
 pub use mock_core::*;
-pub mod mock_fast_add_peer;
-pub use mock_fast_add_peer::*;
 pub mod mock_engine_store_server;
+pub mod mock_fast_add_peer;
 pub use mock_engine_store_server::*;
 pub mod mock_page_storage;
 pub use mock_page_storage::*;
+pub mod mock_ffi;
+pub use mock_ffi::*;
